+++ packages/lib-cordova/src/index.ts
@@ -16,11 +16,11 @@ export * from '../../lib-shared/src/api/WDONetworkingObjects'
 export * from '../../lib-shared/src/WDOError'
 
 // setup platform specific implementations
-import { WDOCordovaCache, WDONetworkingFactoryCordova, WDOPowerAuthFactoryCordova } from './WDOCordovaPlatform'
+import { WDOCordovaCache, WDONetworkingCordovaIntegration, WDOPowerAuthCordovaIntegration } from './WDOCordovaPlatform'
 import { WDOPlatform } from '../../lib-shared/src/WDOPlatform'
 WDOPlatform.cache = new WDOCordovaCache()
-WDOPlatform.networkingFactory = new WDONetworkingFactoryCordova()
-WDOPlatform.powerAuthFactory = new WDOPowerAuthFactoryCordova()
+WDOPlatform.networking = new WDONetworkingCordovaIntegration()
+WDOPlatform.powerAuth = new WDOPowerAuthCordovaIntegration()
 
 // services that depend on platform implementations
 import { WDOBaseActivationService } from '../../lib-shared/src/WDOActivationService'